requirements.txt
torch>=2
torchaudio>=2
faster-whisper==1.0.0
transformers
pandas
setuptools>=65
nltk
uroman-python==1.2.8.1